调整获取流量数据的targetId为null的问题
