+++ api/api_set_traffic_control_traffic_data.go
@@ -1,17 +1,18 @@
 package api
 
 import (
-	"fmt"
 	"github.com/aliyun/alibaba-cloud-sdk-go/services/pairecservice"
 	"github.com/aliyun/aliyun-pairec-config-go-sdk/v2/model"
+	jsoniter "github.com/json-iterator/go"
 )
 
 type TrafficControlTrafficsService service
 
-func (tct *TrafficControlTrafficsService) SetTrafficControlTrafficFData(t model.TrafficControlTaskTrafficData) string {
+func (tct *TrafficControlTrafficsService) SetTrafficControlTrafficFData(t model.TrafficControlTaskTrafficData) (string, error) {
 	t.InstanceId = tct.instanceId
 	req := pairecservice.CreateUpdateTrafficControlTaskTrafficRequest()
-	body := fmt.Sprintf("%v", t)
+	body, _ := jsoniter.MarshalToString(t)
+
 	req.TrafficControlTaskId = t.TrafficControlTaskId
 	req.Body = body
 	req.SetDomain(tct.client.GetDomain())
@@ -19,9 +20,8 @@ func (tct *TrafficControlTrafficsService) SetTrafficControlTrafficFData(t model.
 	response, err := tct.client.UpdateTrafficControlTaskTraffic(req)
 
 	if err != nil {
-		return ""
+		return "", err
 	}
-
-	return response.RequestId
+	return response.RequestId, nil
 
 }
