+++ experiments/client.go
@@ -42,7 +42,7 @@ type ExperimentClient struct {
 	APIClient *api.APIClient
 
 	//
-	sceneMap map[string]*model.Scene
+	SceneMap map[string]*model.Scene
 
 	// sceneParamData map of parameters of scene name
 	sceneParamData map[string]model.SceneParams
@@ -63,7 +63,7 @@ type ExperimentClient struct {
 func NewExperimentClient(instanceId, regionId, accessKeyId, accessKeySecret, environment string, opts ...ClientOption) (*ExperimentClient, error) {
 	client := ExperimentClient{
 		Environment: environment,
-		sceneMap:    make(map[string]*model.Scene, 0),
+		SceneMap:    make(map[string]*model.Scene, 0),
 	}
 
 	var err error
@@ -106,7 +106,7 @@ func (e *ExperimentClient) Validate() error {
 // MatchExperiment specifies to find match experiment by the ExperimentContext
 // If not find the scene return error or return ExperimentResult
 func (e *ExperimentClient) MatchExperiment(sceneName string, experimentContext *model.ExperimentContext) *model.ExperimentResult {
-	sceneData := e.sceneMap
+	sceneData := e.SceneMap
 
 	scene, exist := sceneData[sceneName]
